requirements.txt
google-cloud-bigquery==3.25.0
google-auth==2.29.0